接口全部对接完成，明日开始备案提审

+++ src/main/java/com/mcf/relationship/infra/model/RelationshipDO.java
@@ -77,7 +77,7 @@ public class RelationshipDO implements Serializable {
     /**
      * 是否已删除
      */
-    private Long deleted;
+    private Integer deleted;
 
     /**
      * 创建时间
